Moving http_state_manager_test out of integration
While it hits multiple packages, it's not quite high enough to count as
integration IMO
