+++ test/empty.go
@@ -1,0 +1,4 @@
+package test
+
+// This package exists just to have a place to have multiple-package tests to
+// simplify import issues
